basic room code support
 * ✅ 1. starting screen / has a place to start hosting a room => /admin
 * ✅ 2. when registering the room, store user's web socket connections in room object
 * ✅ 3. keep track of who is host
 * ✅ 4. keep track of each ws connection matched to a user
 * ✅ 5. only broadcast to those connected to the room
